requirements.txt
certifi==2022.12.7
pip==22.3.1
python-dotenv==0.21.1
setuptools==65.6.3
wheel==0.37.1
